Updated chord chart with href on node click
