+++ http_src/vue/ntop_vue.js
@@ -94,7 +94,6 @@ import { default as PageNetworksList } from "./page-networks-list.vue"
 import { default as PageDHCPLeases } from "./page-DHCP-leases.vue"
 import { default as PageExporterInterfaceOverview } from "./page-exporter-interface-overview.vue"
 import { default as PageExporterMap } from "./page-exporter-map.vue"
-import { default as PageChordTest } from "./page-chord-test.vue"
 
 /* Testing page */
 import { default as PageTest } from "./page-test.vue";
@@ -217,7 +216,6 @@ let ntopVue = {
     PageDHCPLeases: PageDHCPLeases,
     PageExporterInterfaceOverview: PageExporterInterfaceOverview,
     PageExporterMap: PageExporterMap,
-    PageChordTest: PageChordTest,
 
     /* Testing page */
     PageTest: PageTest,
